--- routers/menuCategoryRouter.router.js
@@ -1,7 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const menuCategoryController = require('../controllers/menuCategory.controller');
-
-router.get('/', menuCategoryController.menuCategory);
-
-module.exports = router;
