--- services/menuCategory.service.js
@@ -1,1 +1,0 @@
-// 요구하는 서비스 X
